+++ src/forr/input.rs
@@ -1,4 +1,4 @@
-use std::ops::Range;
+use std::ops::{Bound, Range};
 
 use proc_macro2::{Span, TokenStream, TokenTree};
 use proc_macro_utils::TokenTreePunct;
@@ -13,6 +13,7 @@ pub struct MacroInput {
     pub body: TokenStream,
 }
 
+#[derive(Clone, Debug)]
 pub enum Vars {
     Single(Var),
     Plain(Vec<Vars>),
@@ -24,7 +25,30 @@ pub enum Vars {
     },
 }
 
-#[derive(Clone)]
+impl Vars {
+    #[allow(clippy::type_complexity)]
+    pub fn split_tuple(&self) -> Option<((Var, (Bound<usize>, Bound<usize>)), Vars)> {
+        let mut tuple = None;
+        let mut meta = Vec::new();
+
+        let Self::Plain(vars) = &self else { unreachable!() };
+
+        for var in vars {
+            match var {
+                Vars::Function {
+                    function: VarsFunction::Tuples(var, bounds),
+                    ..
+                } => tuple = Some((var.clone(), *bounds)),
+                vars @ Vars::Single(var) if var.typ.is_meta() => meta.push(vars.clone()),
+                _ => return None,
+            }
+        }
+
+        tuple.map(|tuple| (tuple, Vars::Plain(meta)))
+    }
+}
+
+#[derive(Clone, Debug)]
 pub struct Var {
     pub name: Ident,
     pub typ: Typ,
@@ -32,7 +56,7 @@ pub struct Var {
     pub span: Range<Span>,
 }
 
-#[derive(Clone, Copy, PartialEq)]
+#[derive(Clone, Copy, PartialEq, Debug)]
 pub enum Typ {
     Expr,
     Ty,
@@ -46,8 +70,10 @@ pub enum Typ {
     Literal,
 }
 
+#[derive(Clone, Debug)]
 pub enum VarsFunction {
     Casing(Vec<(Ident, Casing)>),
+    Tuples(Var, (Bound<usize>, Bound<usize>)),
 }
 
 #[allow(non_camel_case_types)]
@@ -69,13 +95,13 @@ pub enum ValuesFunction {
     Ident(Ident, usize),
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub enum Prefix {
     Pound,
     Dollar,
 }
 impl Prefix {
-    pub fn test(&self) -> fn(&TokenTree) -> bool {
+    pub fn test(self) -> fn(&TokenTree) -> bool {
         match self {
             Prefix::Pound => TokenTree::is_pound,
             Prefix::Dollar => TokenTree::is_dollar,